modify the router and the making register and login pages
