+++ src/App.js
@@ -1,18 +1,15 @@
 import React from 'react';
-
 import './App.css';
 import { BrowserRouter as Router,Route} from 'react-router-dom';
 
-
 class App extends React.Component {
-render(){
-return(
-<Router >
+    render(){
+        return(
+            <Router >
                 <div>
-	</div>
-</Router>
-)
-}
+                </div>
+            </Router>
+        )
+    }
 }
 export default App;
-
