--- src/components/login.js
@@ -1,13 +1,0 @@
-import React from 'react';
- 
-class Login extends React.Component{
-render(){
-return(
-<div>
-<div>登录界面</div>
-</div>
-);
-}
-}
- 
-export default Login;
