--- src/components/register.js
@@ -1,16 +1,0 @@
-import React from 'react';
-import ReactDOM from 'react-dom';
-import { Layout } from 'antd';
-
- 
-class register extends React.Component{
-render(){
-return(
-<div>
-<div>注册界面</div>
-</div>
-);
-}
-}
- 
-export default register;
